% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsGadget.R
\docType{data}
\name{extraOutputs}
\alias{extraOutputs}
\title{Model outputs}
\format{
An object of class \code{character} of length 52.
}
\usage{
extraOutputs
}
\description{
Model outputs
}
\keyword{datasets}
