% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetCode.R
\name{gadgetCode}
\alias{gadgetCode}
\title{Graph Output Server Function}
\usage{
gadgetCode(input, output, session)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny Outputs.}

\item{session}{Session object.}
}
\value{
list with following components
\describe{
  \item{xvar}{reactive character string indicating x variable selection}
  \item{yvar}{reactive character string indicating y variable selection}
}
}
\description{
Graph Output Server Function
}
