% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetCode.R
\name{gadgetCodeUI}
\alias{gadgetCodeUI}
\title{Patient Network UI Function}
\usage{
gadgetCodeUI(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Patient Network UI Function
}
