% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{generate_network}
\alias{generate_network}
\title{Function to create a visNetwork object in the 'Human Overview' Panel}
\usage{
generate_network(nodes, edges, usephysics = FALSE)
}
\arguments{
\item{nodes}{Network nodes. A dataframe.}

\item{edges}{Network edges. A dataframe.}

\item{usephysics}{Whether to use visNetwork physics. FALSE by default.}
}
\value{
a visNetwork object.
}
\description{
Function to create a visNetwork object in the 'Human Overview' Panel
}
