% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{generate_nodes_pat}
\alias{generate_nodes_pat}
\title{Function to create nodes for a patient}
\usage{
generate_nodes_pat()
}
\value{
a dataframe containing all nodes
}
\description{
Function to create nodes for a patient
}
