% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsSimulations.R
\name{graphSimulationsUI}
\alias{graphSimulationsUI}
\title{Graph Simulations (Hyper/Hypocapnia) Output UI Function}
\usage{
graphSimulationsUI(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Graph Simulations (Hyper/Hypocapnia) Output UI Function
}
