% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{lexicon}
\alias{lexicon}
\title{Lexicon Server Function}
\usage{
lexicon(input, output, session)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny Outputs.}

\item{session}{Session object.}
}
\value{
list with following components
}
\description{
Lexicon Server Function
}
