% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{lexiconUI}
\alias{lexiconUI}
\title{Lexicon Page}
\usage{
lexiconUI(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Lexicon Page
}
