% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsGadget.R
\docType{data}
\name{pars_dropdown_names}
\alias{pars_dropdown_names}
\title{Model outputs for bottom right box}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
pars_dropdown_names
}
\description{
Model outputs for bottom right box
}
\keyword{datasets}
