% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientNetwork.R
\name{patientNetwork}
\alias{patientNetwork}
\title{Patient Network Server Function}
\usage{
patientNetwork(input, output, session)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny Outputs.}

\item{session}{Session object.}
}
\value{
list with following components
\describe{
  \item{xvar}{reactive character string indicating x variable selection}
  \item{yvar}{reactive character string indicating y variable selection}
}
}
\description{
Patient Network Server Function
}
