% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientNetwork.R
\name{patientNetworkUI}
\alias{patientNetworkUI}
\title{Patient Network UI Function (Human Overview)}
\usage{
patientNetworkUI(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tag}} containing UI elements
}
\description{
Patient Network UI Function (Human Overview)
}
