% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsGadget.R
\docType{data}
\name{plot_options_params}
\alias{plot_options_params}
\title{Plot options parameters}
\format{
An object of class \code{character} of length 25.
}
\usage{
plot_options_params
}
\description{
Plot options parameters
}
\keyword{datasets}
