% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsGadget.R
\docType{data}
\name{state_gadget}
\alias{state_gadget}
\title{initial values for the gadget}
\format{
An object of class \code{numeric} of length 32.
}
\usage{
state_gadget
}
\description{
initial values for the gadget
}
\keyword{datasets}
