% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamCards.R
\name{teamCard}
\alias{teamCard}
\title{Team Card Server module}
\usage{
teamCard(input, output, session, src, name, position)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny Outputs.}

\item{session}{Session object.}

\item{src}{Path to the profile image.}

\item{name}{Staff member name.}

\item{position}{Staff member current position}
}
\value{
a card containing people's information
}
\description{
Team Card Server module
}
