% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamCards.R
\name{teamCardUI}
\alias{teamCardUI}
\title{Team Card UI module}
\usage{
teamCardUI(id, width = 4)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{width}{Card width. 4 by default.}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Team Card UI module
}
